[AS7-6926] Verify javax.xml.stream usage for bundles
