+++ testsuite/integration/osgi/src/test/java/org/jboss/as/test/integration/osgi/xml/StreamReaderTestCase.java
@@ -19,14 +19,15 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.osgi.jaxp;
+package org.jboss.as.test.integration.osgi.xml;
 
-import static org.junit.Assert.assertEquals;
 import java.io.InputStream;
 import java.net.URL;
 
-import javax.xml.parsers.SAXParser;
-import javax.xml.parsers.SAXParserFactory;
+import javax.xml.stream.XMLEventReader;
+import javax.xml.stream.XMLInputFactory;
+import javax.xml.stream.events.Characters;
+import javax.xml.stream.events.XMLEvent;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
@@ -36,40 +37,38 @@ import org.jboss.osgi.metadata.OSGiManifestBuilder;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.Asset;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
+import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
-import org.osgi.util.tracker.ServiceTracker;
-import org.xml.sax.SAXException;
-import org.xml.sax.helpers.DefaultHandler;
 
 /**
- * A test that uses a SAX parser to read an XML document.
+ * A test that uses the javax.xml.stream API to read an XML document.
  *
  * @author thomas.diesler@jboss.com
- * @since 21-Jul-2009
+ * @since 17-Apr-2013
  */
 @RunWith(Arquillian.class)
-public class SAXParserTestCase {
+public class StreamReaderTestCase {
 
     @ArquillianResource
     BundleContext context;
+
     @ArquillianResource
     Bundle bundle;
 
     @Deployment
     public static JavaArchive createdeployment() {
-        final JavaArchive archive = ShrinkWrap.create(JavaArchive.class, "sax-parser.jar");
+        final JavaArchive archive = ShrinkWrap.create(JavaArchive.class, "stream-reader.jar");
         archive.addClasses(FrameworkUtils.class);
-        archive.addAsResource(SAXParserTestCase.class.getPackage(), "simple.xml", "simple.xml");
+        archive.addAsResource(StreamReaderTestCase.class.getPackage(), "simple.xml", "simple.xml");
         archive.setManifest(new Asset() {
             public InputStream openStream() {
                 OSGiManifestBuilder builder = OSGiManifestBuilder.newInstance();
                 builder.addBundleSymbolicName(archive.getName());
                 builder.addBundleManifestVersion(2);
-                builder.addImportPackages(SAXParser.class, SAXException.class, DefaultHandler.class);
-                builder.addImportPackages(ServiceTracker.class);
+                builder.addImportPackages(XMLEventReader.class, XMLEvent.class);
                 return builder.openStream();
             }
         });
@@ -78,38 +77,17 @@ public class SAXParserTestCase {
 
     @Test
     public void testSAXParserFactoryAPI() throws Exception {
-        SAXParserFactory factory = SAXParserFactory.newInstance();
-        parse(factory);
-    }
-
-    @Test
-    public void testSAXParserFactoryService() throws Exception {
-        SAXParserFactory factory = FrameworkUtils.waitForService(context, SAXParserFactory.class);
-        parse(factory);
-    }
-
-    private void parse(SAXParserFactory factory) throws Exception {
-        factory.setNamespaceAware(true);
-        factory.setValidating(false);
-
-        SAXParser saxParser = factory.newSAXParser();
+        String content = null;
         URL resURL = bundle.getResource("simple.xml");
-
-        SAXHandler saxHandler = new SAXHandler();
-        saxParser.parse(resURL.openStream(), saxHandler);
-        assertEquals("content", saxHandler.getContent());
-    }
-
-    static class SAXHandler extends DefaultHandler {
-        private String content;
-
-        @Override
-        public void characters(char[] ch, int start, int length) throws SAXException {
-            content = new String(ch, start, length);
+        XMLInputFactory factory = XMLInputFactory.newInstance();
+        XMLEventReader reader = factory.createXMLEventReader(resURL.openStream());
+        while(reader.hasNext()) {
+            XMLEvent event = reader.nextEvent();
+            if (event.isCharacters()) {
+                Characters chars = event.asCharacters();
+                content = chars.getData();
+            }
         }
-
-        public String getContent() {
-            return content;
-        }
+        Assert.assertEquals("content", content);
     }
 }
