+++ testsuite/integration/osgi/src/test/java/org/jboss/as/test/integration/osgi/xml/simple.xml
